% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_binocular.R
\docType{data}
\name{data_binocular}
\alias{data_binocular}
\title{Example raw eye position data  (Eyelink 1000), recorded binocularly.}
\format{
A data frame with 4,039 rows and 7 variables:
\describe{
\item{time}{Time at which eye sample was taken (internal to Eyelink tracker)}
\item{x_left}{x position of the LEFT eye recorded by the Eye-tracker}
\item{y_left}{y position of the LEFT eye recorded by the Eye-tracker}
\item{pupil_left}{Pupil diameter of the LEFT eye in pixels}
\item{x_right}{x position of the RIGHT eye recorded by the Eye-tracker}
\item{y_right}{y position of the RIGHT eye recorded by the Eye-tracker}
\item{pupil_right}{Pupil diameter of the RIGHT eye in pixels}
}
}
\usage{
data_binocular
}
\description{
This data contains a subset of the data of a participant reading a passage.
}
\keyword{datasets}
