% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDT.R
\name{IDT}
\alias{IDT}
\title{Detect fixations or saccades using the Identification by Dispersion Threshold  (I-DT) method}
\usage{
IDT(
  data,
  dva_x = 0.0187,
  dva_y = 0.0192,
  window_threshold = 50,
  disp_threshold = 1,
  return_saccades = FALSE,
  min_sacc_amplitude = 0.15
)
}
\arguments{
\item{data}{A data frame containing the raw data samples.
Must include at least the following columns: "time", "x", and "y".}

\item{dva_x}{Degrees of visual angle per pixel (x dimension)}

\item{dva_y}{Degrees of visual angle per pixel (y dimension)}

\item{window_threshold}{Minimum fixation duration window threshold in milliseconds (I-DT algorithm)}

\item{disp_threshold}{Minimum sample dispersion threshold in visual angle (I-DT algorithm)}

\item{return_saccades}{Logical. If \code{FALSE} (default), return fixations.
If \code{TRUE}, return saccades instead.

If \code{return_saccades = FALSE}, the function returns fixation-level data
with the following columns:
\describe{
\item{\code{fix_id}}{Sequential fixation identifier.}
\item{\code{fix_start}}{Start time of the fixation, in the same units as \code{data$time}.}
\item{\code{fix_end}}{End time of the fixation, in the same units as \code{data$time}.}
\item{\code{x}}{Mean horizontal gaze position during the fixation, in pixels.}
\item{\code{y}}{Mean vertical gaze position during the fixation, in pixels.}
\item{\code{fix_dur}}{Fixation duration, computed as \code{fix_end - fix_start}.}
\item{\code{avg_velocity}}{Mean sample-to-sample velocity during the fixation, in degrees/second.}
\item{\code{n_samples}}{Number of samples contributing to the fixation.}
}

If \code{return_saccades = TRUE}, the function returns saccade-level data
with the following columns:
\describe{
\item{\code{sacc_id}}{Sequential event identifier for the detected saccade.}
\item{\code{sacc_start}}{Start time of the saccade, in the same units as \code{data$time}.}
\item{\code{sacc_end}}{End time of the saccade, in the same units as \code{data$time}.}
\item{\code{sacc_dur}}{Saccade duration, computed as \code{sacc_end - sacc_start}.}
\item{\code{x_start}}{Horizontal gaze position at saccade onset, in pixels.}
\item{\code{y_start}}{Vertical gaze position at saccade onset, in pixels.}
\item{\code{x_end}}{Horizontal gaze position at saccade offset, in pixels.}
\item{\code{y_end}}{Vertical gaze position at saccade offset, in pixels.}
\item{\code{sacc_amplitude_deg}}{Saccade amplitude in degrees of visual angle, computed from the horizontal and vertical displacement between saccade onset and offset.}
\item{\code{avg_velocity_deg}}{Mean sample-to-sample velocity during the saccade, in degrees/second.}
\item{\code{peak_velocity_deg}}{Maximum sample-to-sample velocity during the saccade, in degrees/second.}
\item{\code{n_samples}}{Number of samples contributing to the saccade.}
}}

\item{min_sacc_amplitude:}{Minimum saccade amplitude in degrees (if using saccade detection). By default, this is set to 0.15.}
}
\description{
This function detects fixations from raw eye position samples
using the Identification by Dispersion Threshold  (I-DT). Note that the algorithm
requires calculations based on visual angle (see
\link[=VisualAngle]{VisualAngle}). As a result, the degrees of visual
angle per pixel need to be provided for the x- and y-dimensions (
\code{dva_x} and \code{dva_y}). While saccade data can be outputted, note that this algorithm is not optimised for saccade detection.
}
\references{
Salvucci, D. D., & Goldberg, J. H. (2000).
Identifying fixations and saccades in eye-tracking protocols.
In \emph{Proceedings of the 2000 Symposium on Eye Tracking
Research & Applications} (pp. 71--78).
New York: ACM Press.
\doi{10.1145/355017.355028}
}
\author{
Martin R. Vasilev, Yixin Ding
}
