% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IHMM.R
\name{IHMM}
\alias{IHMM}
\title{Detect fixations or saccades using Identification by Hidden Markov Models}
\usage{
IHMM(
  data,
  dva_x = NULL,
  dva_y = NULL,
  return_saccades = FALSE,
  min_fix_dur = 50,
  min_sacc_amplitude = 0.15,
  nstates = 2,
  return_samples = FALSE
)
}
\arguments{
\item{data}{A data frame containing the raw data samples.
Must include at least the following columns: \code{"time"},
\code{"x"}, and \code{"y"}. Time must be measured in milliseconds.}

\item{dva_x}{Degrees of visual angle per pixel (x dimension);
optional, but recommended.}

\item{dva_y}{Degrees of visual angle per pixel (y dimension);
optional, but recommended.}

\item{return_saccades}{Logical. If \code{FALSE} (default), return
fixations. If \code{TRUE}, return saccades instead.}

\item{min_fix_dur}{Minimum fixation duration (in ms) to extract.
Fixations shorter than this value are discarded.}

\item{min_sacc_amplitude}{Minimum saccade amplitude in degrees
(if using saccade detection and degrees of visual angle are supplied).
By default, this is set to 0.15 deg.}

\item{nstates}{Number of hidden states in the HMM.
By default, this is set to 2 (fixation and saccade states).}

\item{return_samples}{A logical indicating whether to return raw samples (with states mapped). FALSE by default, can be set to TRUE to obtain the samples instead.}
}
\value{
A data frame with one row per detected event.
The columns depend on whether fixations or saccades are returned.

If \code{return_saccades = FALSE}, the function returns fixation-level data
with the following columns:

\describe{
\item{\code{fix_id}}{Sequential fixation identifier.}
\item{\code{fix_start}}{Start time of the fixation, in the same units as \code{data$time}.}
\item{\code{fix_end}}{End time of the fixation, in the same units as \code{data$time}.}
\item{\code{fix_dur}}{Fixation duration, computed as \code{fix_end - fix_start}.}
\item{\code{x}}{Mean horizontal gaze position during the fixation, in pixels.}
\item{\code{y}}{Mean vertical gaze position during the fixation, in pixels.}
\item{\code{avg_velocity}}{Mean sample-to-sample velocity during the fixation.}
\item{\code{n_samples}}{Number of samples contributing to the fixation.}
}

If \code{return_saccades = TRUE}, the function returns saccade-level data
with the following columns:

\describe{
\item{\code{sacc_id}}{Sequential event identifier for the detected saccade.}
\item{\code{sacc_start}}{Start time of the saccade, in the same units as \code{data$time}.}
\item{\code{sacc_end}}{End time of the saccade, in the same units as \code{data$time}.}
\item{\code{sacc_dur}}{Saccade duration, computed as \code{sacc_end - sacc_start}.}
\item{\code{x_start}}{Horizontal gaze position at saccade onset, in pixels.}
\item{\code{y_start}}{Vertical gaze position at saccade onset, in pixels.}
\item{\code{x_end}}{Horizontal gaze position at saccade offset, in pixels.}
\item{\code{y_end}}{Vertical gaze position at saccade offset, in pixels.}
\item{\code{sacc_amplitude_deg}}{Saccade amplitude in degrees of visual angle (if degrees of visual angle are supplied).}
\item{\code{avg_velocity_deg}}{Mean sample-to-sample velocity during the saccade, in degrees/second (if degrees of visual angle are supplied).}
\item{\code{peak_velocity_deg}}{Maximum sample-to-sample velocity during the saccade, in degrees/second (if degrees of visual angle are supplied).}
\item{\code{n_samples}}{Number of samples contributing to the saccade.}
}
}
\description{
Detects fixations or saccades from eye-tracking samples using a Hidden
Markov Model (HMM) approach inspired by the Identification by Hidden
Markov Models (I-HMM) method proposed by Salvucci and Goldberg (2000).
}
\details{
The algorithm estimates hidden fixation and saccade states from
sample-to-sample gaze velocities using a Gaussian Hidden Markov Model
fitted with the \pkg{depmixS4} package. Velocities are calculated from
successive gaze samples and the Viterbi algorithm is used to infer the
most likely sequence of hidden states.

By default, the function returns fixation data.
To return saccade data instead, set \code{return_saccades = TRUE}.

Note that the function works without degrees of visual angle being supplied.
However, we strongly recommend that \code{dva_x} and \code{dva_y} are
provided (especially if saccades are of interest) so that velocities and
amplitudes can be returned in deg/s and deg, respectively (see
\link[=VisualAngle]{VisualAngle}).
}
\references{
Salvucci, D. D., & Goldberg, J. H. (2000).
Identifying fixations and saccades in eye-tracking protocols.
\emph{Proceedings of the 2000 Symposium on Eye Tracking Research &
Applications}, 71--78.
}
\seealso{
\link[=IVT]{IVT},
\link[=IDT]{IDT},
\link[=EngbertKliegl03]{EngbertKliegl03},
\link[=VisualAngle]{VisualAngle}
}
\author{
Martin R. Vasilev
}
