% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManualEventLabeler.R
\name{ManualEventLabeler}
\alias{ManualEventLabeler}
\title{Launches a manual event labeler}
\usage{
ManualEventLabeler()
}
\value{
Invisibly returns the path to the HTML app.
}
\description{
Opens a browser-based manual labeler for saccadic events.
}
