% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smooth.R
\name{SmoothSamples}
\alias{SmoothSamples}
\title{Smooth eye-tracking samples using different methods}
\usage{
SmoothSamples(
  data,
  method = "Mean",
  window_size = 5,
  align = "center",
  p = 5,
  n = 23
)
}
\arguments{
\item{data}{A data frame of eye sample to smooth. Must include at least the following columns: "x", and "y".}

\item{method}{Data smooting method. Possible options are: "Mean", "Median", "SG" for Savitzky-Golay filter.}

\item{window_size}{Moving window size (number of samples) for moving average/ median methods. Must be a positive odd integer.}

\item{align}{Alignment of the window: "center" (default), "left", or "right".}

\item{p}{Savitzky-Golay filter order (default= 5)}

\item{n}{Savitzky-Golay filter length (must be odd, default= 23).}
}
\value{
A numeric vector of the same length as \code{data}, containing the smoothed values.
}
\description{
This function performs smoothing of eye samples prior to event detection to reduce noise. Samples can be smoothed using different methods, such as moving average, moving median, Savitzky-Golay filter.
}
\examples{
Oz_smooth <- SmoothSamples(data_Oz, method= "Mean")
Oz_smooth <- SmoothSamples(data_Oz, method= "Median")

}
