% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetectFixations.R
\name{DetectFixations}
\alias{DetectFixations}
\title{Detect fixations from raw eye sample data}
\usage{
DetectFixations(
  data,
  method = "I-DT",
  dva_x = 0.0187,
  dva_y = 0.0192,
  window_threshold = 100,
  disp_threshold = 1
)
}
\arguments{
\item{data}{A data frame containing the raw data samples.
Must include at least the following columns: "time", "x", and "y".}

\item{method}{The fixation detection algorithm to use (I-DT)}

\item{dva_x}{Degrees of visual angle per pixel (x dimension)}

\item{dva_y}{Degrees of visual angle per pixel (y dimension)}

\item{window_threshold}{Minimum fixation duration window threshold in milliseconds (I-DT algorithm)}

\item{disp_threshold}{Minimum sample disperson threshold in visual angle (I-DT algorithm)}
}
\value{
A data frame with one row per detected fixation and the following columns:
\describe{
\item{fix_id}{Sequential fixation identifier within the trial.}
\item{start_fix}{Start time of the fixation (in milliseconds, eye-tracker internal clock).}
\item{end_fix}{End time of the fixation (in milliseconds).}
\item{x}{Mean horizontal gaze position of all samples in the fixation (in pixels).}
\item{y}{Mean vertical gaze position of all samples in the fixation (in pixels).}
\item{FixDur}{Fixation duration in milliseconds, computed as
\code{end_fix - start_fix}.}
\item{dispersion_deg}{Total spatial dispersion of samples in the fixation,
computed as the sum of horizontal and vertical dispersion in degrees of visual angle
(I-DT algorithm).}
\item{n_samples}{Number of raw eye-tracking samples contributing to the fixation.}
}
}
\description{
Detect fixations from raw eye sample data
}
\examples{
data("data_Oz")

fixations <- DetectFixations(
  data = data_Oz, method = "I-DT",
  dva_x = 0.0187, dva_y = 0.0192,
  window_threshold = 100, disp_threshold = 1
)

head(fixations)

}
\author{
Martin R. Vasilev, Yixin Ding
}
